refactor(watch): split pure camera/horizon math into WatchModeController.Math.cs
